+++ Coimbra/CancellationTokenSourceUtility.cs
@@ -1,5 +1,6 @@
 ﻿#nullable enable
 
+using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Threading;