+++ Coimbra/Disposable.cs
@@ -7,14 +7,14 @@ namespace Coimbra
     /// Create a disposable from any type.
     /// </summary>
     [Preserve]
-    public readonly ref struct Disposable<T>
+    public ref struct Disposable<T>
     {
         public delegate void DisposeHandler([CanBeNull] in T value);
 
         [CanBeNull]
         public readonly T Value;
 
-        private readonly DisposeHandler _onDispose;
+        private DisposeHandler _onDispose;
 
         public Disposable([CanBeNull] in T value, [CanBeNull] DisposeHandler onDispose)
         {
@@ -25,6 +25,7 @@ namespace Coimbra
         public void Dispose()
         {
             _onDispose?.Invoke(in Value);
+            _onDispose = null;
         }
     }
 }